+++ lib/dnsimple/domains.ex
@@ -129,28 +129,6 @@ defmodule Dnsimple.Domains do
 
 
   @doc """
-  Resets the domain API token used for authentication in APIv1.
-
-  See:
-  - https://developer.dnsimple.com/v2/domains/#reset-token
-
-  ## Examples:
-
-      client = %Dnsimple.Client{access_token: "a1b2c3d4"}
-      {:ok, response} = Dnsimple.Domains.reset_domain_token(client, account_id = 1010, domain_id = 123)
-      {:ok, response} = Dnsimple.Domains.reset_domain_token(client, account_id = 1010, domain_id = "example.io")
-
-  """
-  @spec reset_domain_token(Client.t, String.t | integer, String.t | integer, Keyword.t) :: {:ok|:error, Response.t}
-  def reset_domain_token(client, account_id, domain_id, options \\ []) do
-    url = Client.versioned("/#{account_id}/domains/#{domain_id}/token")
-
-    Client.post(client, url, Client.empty_body(), options)
-    |> Response.parse(%{"data" => %Domain{}})
-  end
-
-
-  @doc """
   Enable DNSSEC for the domain in the account.
 
   See: