+++ CHANGELOG.md
@@ -1,9 +1,7 @@
 # Changelog
 
 ## main
-- CHANGED: Bump httpoison to 1.8.0 (dnsimple/dnsimple-elixir#148)
-- CHANGED: Bump exdoc to 0.23.0 (dnsimple/dnsimple-elixir#146)
-- CHANGED: Bump exvcr to 0.12.1 (dnsimple/dnsimple-elixir#145)
+
 - CHANGED: Fix warning about ExvcrUtils (dnsimple/dnsimple-elixir#139)
 - CHANGED: `Dnsimple.Domain` struct now has `expires_at` (timestamp) to be used in favor of `expires_on` (date only).
   (dnsimple/dnsimple-elixir#135)